+++ src/Ext/WeaponType/Body.h
@@ -38,6 +38,8 @@ public:
 		Nullable<int> Strafing_EndDelay;
 		Valueable<AffectedTarget> CanTarget;
 		Valueable<AffectedHouse> CanTargetHouses;
+		Valueable<double> CanTarget_MaxHealth;
+		Valueable<double> CanTarget_MinHealth;
 		ValueableVector<int> Burst_Delays;
 		Valueable<bool> Burst_FireWithinSequence;
 		Valueable<AreaFireTarget> AreaFire_Target;
@@ -87,6 +89,8 @@ public:
 		Valueable<int> LaserThickness;
 		Valueable<bool> Laser_IsTracking;
 
+		bool SkipWeaponPicking;
+
 		ExtData(WeaponTypeClass* OwnerObject) : Extension<WeaponTypeClass>(OwnerObject)
 			, DiskLaser_Radius { DiskLaserClass::Radius }
 			, ProjectileRange { Leptons(100000) }
@@ -104,6 +108,8 @@ public:
 			, Strafing_EndDelay {}
 			, CanTarget { AffectedTarget::All }
 			, CanTargetHouses { AffectedHouse::All }
+			, CanTarget_MaxHealth { 1.0 }
+			, CanTarget_MinHealth { 0.0 }
 			, Burst_Delays {}
 			, Burst_FireWithinSequence { false }
 			, AreaFire_Target { AreaFireTarget::Base }
@@ -151,11 +157,14 @@ public:
 			, Beam_IsHouseColor { false }
 			, LaserThickness { 3 }
 			, Laser_IsTracking { false }
+			, SkipWeaponPicking { true }
 		{ }
 
 		int GetBurstDelay(int burstIndex) const;
 
 		bool HasRequiredAttachedEffects(TechnoClass* pTechno, TechnoClass* pFirer) const;
+
+		bool IsHealthRatioEligible(TechnoClass* const pTarget) const;
 
 		virtual ~ExtData() = default;